worker/pyproject.toml
[project]
name = "memexflow-worker"
version = "0.1.0"
description = "MemexFlow AI Worker - ingestion, extraction, briefing, retrieval"
requires-python = ">=3.11"
dependencies = [
    "supabase>=2.0.0",
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "structlog>=24.0.0",
    "tenacity>=8.0.0",
]

[project.scripts]
memexflow-worker = "worker.main:run"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.6.0",
    "mypy>=1.10.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
